refactor(Cantines): API: bouger la logique dédiée aux gestionnaires dans un nouveau fichier canteen_managers.py
