+++ api/tests/test_canteen_published.py
@@ -1134,99 +1134,6 @@ class PublishedCanteenDetailApiTest(APITestCase):
         self.assertEqual(len(body.get("approDiagnostics")), 1)
 
 
-class TestPublishedCanteenClaimApiTest(APITestCase):
-    @authenticate
-    def test_canteen_claim_request(self):
-        canteen = CanteenFactory()
-        canteen.managers.clear()
-
-        response = self.client.post(reverse("claim_canteen", kwargs={"canteen_pk": canteen.id}), None)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        body = response.json()
-        self.assertEqual(body["id"], canteen.id)
-        self.assertEqual(body["name"], canteen.name)
-        user = authenticate.user
-        self.assertEqual(canteen.managers.first().id, user.id)
-        self.assertEqual(canteen.managers.count(), 1)
-        canteen.refresh_from_db()
-        self.assertEqual(canteen.claimed_by, user)
-        self.assertTrue(canteen.has_been_claimed)
-
-    @authenticate
-    def test_canteen_not_filled_can_be_claimed(self):
-        canteen = CanteenFactory()
-        canteen.managers.clear()
-        canteen.siret = None
-        canteen.save(skip_validations=True)
-
-        self.assertIsNone(canteen.siret)
-        self.assertFalse(canteen.is_filled)
-
-        response = self.client.post(reverse("claim_canteen", kwargs={"canteen_pk": canteen.id}), None)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        body = response.json()
-        self.assertEqual(body["id"], canteen.id)
-        self.assertEqual(body["name"], canteen.name)
-        user = authenticate.user
-        self.assertEqual(canteen.managers.first().id, user.id)
-        self.assertEqual(canteen.managers.count(), 1)
-        canteen.refresh_from_db()
-        self.assertEqual(canteen.claimed_by, user)
-        self.assertTrue(canteen.has_been_claimed)
-
-    @authenticate
-    def test_canteen_claim_request_fails_when_already_claimed(self):
-        canteen = CanteenFactory()
-        self.assertGreater(canteen.managers.count(), 0)
-        user = authenticate.user
-        self.assertFalse(canteen.managers.filter(id=user.id).exists())
-
-        response = self.client.post(reverse("claim_canteen", kwargs={"canteen_pk": canteen.id}), None)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertFalse(canteen.managers.filter(id=user.id).exists())
-        canteen.refresh_from_db()
-        self.assertFalse(canteen.has_been_claimed)
-
-    @authenticate
-    def test_undo_claim_canteen(self):
-        canteen = CanteenFactory(claimed_by=authenticate.user, has_been_claimed=True, managers=[authenticate.user])
-
-        response = self.client.post(reverse("undo_claim_canteen", kwargs={"canteen_pk": canteen.id}), None)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertFalse(canteen.managers.filter(id=authenticate.user.id).exists())
-        canteen.refresh_from_db()
-        self.assertIsNone(canteen.claimed_by)
-        self.assertFalse(canteen.has_been_claimed)
-
-    @authenticate
-    def test_undo_claim_for_canteen_not_filled(self):
-        canteen = CanteenFactory(claimed_by=authenticate.user, has_been_claimed=True, managers=[authenticate.user])
-        canteen.siret = None
-        canteen.save(skip_validations=True)
-
-        self.assertIsNone(canteen.siret)
-        self.assertFalse(canteen.is_filled)
-
-        response = self.client.post(reverse("undo_claim_canteen", kwargs={"canteen_pk": canteen.id}), None)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertFalse(canteen.managers.filter(id=authenticate.user.id).exists())
-        canteen.refresh_from_db()
-        self.assertIsNone(canteen.claimed_by)
-        self.assertFalse(canteen.has_been_claimed)
-
-    @authenticate
-    def test_undo_claim_canteen_fails_if_not_original_claimer(self):
-        other_user = UserFactory()
-        canteen = CanteenFactory(claimed_by=other_user, has_been_claimed=True, managers=[authenticate.user])
-
-        response = self.client.post(reverse("undo_claim_canteen", kwargs={"canteen_pk": canteen.id}), None)
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-        self.assertTrue(canteen.managers.filter(id=authenticate.user.id).exists())
-        canteen.refresh_from_db()
-        self.assertTrue(canteen.has_been_claimed)
-        self.assertEqual(canteen.claimed_by, other_user)
-
-
 class CanteenPreviewDetailApiTest(APITestCase):
     def test_get_single_public_canteen_preview(self):
         """
